:sparkles: Uses cache_file if available
Now that I can cache the op list-items result, I need to use the cache instead
of getting it online, if it exists. It is important to have a TTL on the cache
file too.

On this commit I'm going to add the logic to use the cache file (if applicable)
and also create a rule to use it only when the cache is recent.

Now, I've finished the feature and can start using it.